refactor: 공통 부분 분리, 테스트 분리
- 공통 부분 분리, 테스트 분리
- Embeddable 테스트 분리
